Convert FilterForm to typescript and functional component
