+++ graylog2-web-interface/src/components/event-definitions/event-definition-types/FilterForm.tsx
@@ -1,0 +1,433 @@
+/*
+ * Copyright (C) 2020 Graylog, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+import * as React from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import PropTypes from 'prop-types';
+import camelCase from 'lodash/camelCase';
+import cloneDeep from 'lodash/cloneDeep';
+import debounce from 'lodash/debounce';
+import defaultTo from 'lodash/defaultTo';
+import get from 'lodash/get';
+import isEmpty from 'lodash/isEmpty';
+import merge from 'lodash/merge';
+import memoize from 'lodash/memoize';
+import max from 'lodash/max';
+import union from 'lodash/union';
+import moment from 'moment';
+import { OrderedMap } from 'immutable';
+
+import { useStore } from 'stores/connect';
+import Store from 'logic/local-storage/Store';
+import { MultiSelect, TimeUnitInput, SearchFiltersFormControls } from 'components/common';
+import Query from 'views/logic/queries/Query';
+import Search from 'views/logic/search/Search';
+import { extractDurationAndUnit } from 'components/common/TimeUnitInput';
+import { Alert, ButtonToolbar, ControlLabel, FormGroup, HelpBlock, Input } from 'components/bootstrap';
+import { naturalSortIgnoreCase } from 'util/SortUtils';
+import * as FormsUtils from 'util/FormsUtils';
+import { isPermitted } from 'util/PermissionsMixin';
+import LookupTableParameter from 'views/logic/parameters/LookupTableParameter';
+import { LookupTablesActions, LookupTablesStore } from 'stores/lookup-tables/LookupTablesStore';
+import generateId from 'logic/generateId';
+import parseSearch from 'views/logic/slices/parseSearch';
+import useLocation from 'routing/useLocation';
+import { getPathnameWithoutId } from 'util/URLUtils';
+import { TELEMETRY_EVENT_TYPE } from 'logic/telemetry/Constants';
+import useSendTelemetry from 'logic/telemetry/useSendTelemetry';
+import type User from 'logic/users/User';
+import type { EventDefinition } from 'components/event-definitions/event-definition-types';
+import type { Stream } from 'stores/streams/StreamsStore';
+
+import EditQueryParameterModal from '../event-definition-form/EditQueryParameterModal';
+import commonStyles from '../common/commonStyles.css';
+
+export const PLUGGABLE_CONTROLS_HIDDEN_KEY = 'pluggableSearchBarControlsAreHidden';
+export const TIME_UNITS = ['HOURS', 'MINUTES', 'SECONDS'];
+
+const LOOKUP_PERMISSIONS = [
+  'lookuptables:read',
+];
+
+const buildNewParameter = (name: string) => ({
+  name: name,
+  embryonic: true,
+  type: 'lut-parameter-v1',
+  data_type: 'any',
+  title: 'new title',
+  // has no binding, no need to set binding property
+});
+
+type Props = {
+  currentUser: User,
+  eventDefinition: EventDefinition,
+  onChange: (name: string, config: EventDefinition['config']) => void,
+  streams: Array<Stream>,
+  validation: { errors: { [key: string]: Array<string> } }
+}
+
+const FilterForm = ({
+  currentUser,
+  eventDefinition,
+  onChange,
+  streams,
+  validation,
+} : Props) => {
+  const { execute_every_ms: executeEveryMs, search_within_ms: searchWithinMs } = eventDefinition.config;
+  const searchWithin = extractDurationAndUnit(searchWithinMs, TIME_UNITS);
+  const executeEvery = extractDurationAndUnit(executeEveryMs, TIME_UNITS);
+
+  const { tables } = useStore(LookupTablesStore);
+  const { pathname } = useLocation();
+
+  const sendTelemetry = useSendTelemetry();
+
+  const queryId = generateId();
+  const searchTypeId = generateId();
+
+  const [queryParameterStash, setQueryParameterStash] = useState<object>({});
+  const [searchFiltersHidden, setSearchFiltersHidden] = useState<boolean>(false);
+  const [searchWithinMsUnit, setSearchWithinMsUnit] = useState<string>(searchWithin.unit);
+  const [executeEveryMsUnit, setExecuteEveryMsUnit] = useState<string>(executeEvery.unit);
+  const [searchWithinMsDuration, setSearchWithinMsDuration] = useState<number>(searchWithin.duration);
+  const [executeEveryMsDuration, setExecuteEveryMsDuration] = useState<number>(executeEvery.duration);
+
+  const userCanViewLookupTables = useCallback(
+    () => isPermitted(currentUser.permissions, LOOKUP_PERMISSIONS),
+    [currentUser.permissions],
+  );
+
+  useEffect(() => {
+    if (userCanViewLookupTables()) {
+      LookupTablesActions.searchPaginated(1, 0, undefined, false);
+    }
+  }, [userCanViewLookupTables]);
+
+  const propagateChange = (key, value) => {
+    const config = cloneDeep(eventDefinition.config);
+    config[key] = value;
+    onChange('config', config);
+  };
+
+  const formatStreamIds = memoize(
+    (streamIds) => streamIds
+      .map((streamId) => streams.find((s) => s.id === streamId) || streamId)
+      .map((streamOrId) => {
+        const stream = (typeof streamOrId === 'object' ? streamOrId : { title: streamOrId, id: streamOrId });
+
+        return {
+          label: stream.title,
+          value: stream.id,
+        };
+      })
+      .sort((s1, s2) => naturalSortIgnoreCase(s1.label, s2.label)),
+    (streamIds) => streamIds.join('-'),
+  );
+
+  const syncParamsWithQuery = (paramsInQuery) => {
+    const config = cloneDeep(eventDefinition.config);
+    const queryParameters = config?.query_parameters || [];
+    const keptParameters = [];
+    const staleParameters = {};
+
+    queryParameters.forEach((p) => {
+      if (paramsInQuery.has(p.name)) {
+        keptParameters.push(p);
+      } else {
+        staleParameters[p.name] = p;
+      }
+    });
+
+    const newParameters = [];
+
+    paramsInQuery.forEach((np) => {
+      if (!keptParameters.find((p) => p.name === np)) {
+        if (queryParameterStash[np]) {
+          newParameters.push(queryParameterStash[np]);
+        } else {
+          newParameters.push(buildNewParameter(np));
+        }
+      }
+    });
+
+    setQueryParameterStash(merge(queryParameterStash, staleParameters));
+
+    config.query_parameters = keptParameters.concat(newParameters);
+    onChange('config', config);
+  };
+
+  const parseQuery = debounce((queryString, searchFilters = OrderedMap()) => {
+    if (!userCanViewLookupTables()) {
+      return;
+    }
+
+    const queryBuilder = Query.builder()
+      .id(queryId)
+      .query({ type: 'elasticsearch', query_string: queryString })
+      .timerange({ type: 'relative', range: 1000 })
+      .filters(searchFilters.toList())
+      .searchTypes([{
+        id: searchTypeId,
+        type: 'messages',
+        limit: 10,
+        offset: 0,
+        filter: undefined,
+        filters: undefined,
+        name: undefined,
+        query: undefined,
+        timerange: undefined,
+        streams: undefined,
+        sort: [],
+        decorators: [],
+      }]);
+
+    const query = queryBuilder.build();
+
+    const search = Search.create().toBuilder()
+      .queries([query])
+      .build();
+
+    parseSearch(search).then((res) => {
+      syncParamsWithQuery(res.undeclared);
+    });
+  }, 250);
+
+  const handleConfigChange = (event) => {
+    const { name } = event.target;
+    const value = FormsUtils.getValueFromInput(event.target);
+
+    if (name === '_is_scheduled') {
+      sendTelemetry(TELEMETRY_EVENT_TYPE.EVENTDEFINITION_CONDITION.FILTER_EXECUTED_AUTOMATICALLY_TOGGLED, {
+        app_pathname: getPathnameWithoutId(pathname),
+        app_section: 'event-definition-condition',
+        app_action_value: 'enable-checkbox',
+        is_scheduled: value,
+      });
+    }
+
+    propagateChange(name, value);
+  };
+
+  const handleQueryChange = (event) => {
+    parseQuery(event.target.value);
+    handleConfigChange(event);
+  };
+
+  const handleSearchFiltersChange = (searchFilters) => {
+    const { query } = eventDefinition.config;
+
+    parseQuery(query, searchFilters);
+
+    propagateChange('filters', searchFilters.toArray());
+  };
+
+  const hideFiltersPreview = (value) => {
+    Store.set(PLUGGABLE_CONTROLS_HIDDEN_KEY, value);
+    setSearchFiltersHidden(value);
+  };
+
+  const handleStreamsChange = (nextValue) => {
+    sendTelemetry(TELEMETRY_EVENT_TYPE.EVENTDEFINITION_CONDITION.FILTER_STREAM_SELECTED, {
+      app_pathname: getPathnameWithoutId(pathname),
+      app_section: 'event-definition-condition',
+      app_action_value: 'stream-select',
+    });
+
+    propagateChange('streams', nextValue);
+  };
+
+  const handleTimeRangeChange = (fieldName) => (nextValue, nextUnit) => {
+    if (fieldName === 'search_within_ms' && nextUnit !== searchWithinMsUnit) {
+      sendTelemetry(TELEMETRY_EVENT_TYPE.EVENTDEFINITION_CONDITION.FILTER_SEARCH_WITHIN_THE_LAST_UNIT_CHANGED, {
+        app_pathname: getPathnameWithoutId(pathname),
+        app_section: 'event-definition-condition',
+        app_action_value: 'searchWithinMsUnit-select',
+        new_unit: nextUnit,
+      });
+    } else if (fieldName === 'execute_every_ms' && nextUnit !== executeEveryMsUnit) {
+      sendTelemetry(TELEMETRY_EVENT_TYPE.EVENTDEFINITION_CONDITION.FILTER_EXECUTE_SEARCH_EVERY_UNIT_CHANGED, {
+        app_pathname: getPathnameWithoutId(pathname),
+        app_section: 'event-definition-condition',
+        app_action_value: 'executeEveryMsUnit-select',
+        new_unit: nextUnit,
+      });
+    }
+
+    const durationInMs = moment.duration(max([nextValue, 1]), nextUnit).asMilliseconds();
+
+    propagateChange(fieldName, durationInMs);
+
+    const stateFieldName = camelCase(fieldName);
+
+    if (stateFieldName === 'searchWithinMs') {
+      setSearchWithinMsDuration(nextValue);
+      setSearchWithinMsUnit(nextUnit);
+
+      return;
+    }
+
+    setExecuteEveryMsDuration(nextValue);
+    setExecuteEveryMsUnit(nextUnit);
+  };
+
+  const renderQueryParameters = () => {
+    const queryParameters = eventDefinition?.config?.query_parameters || [];
+
+    const onChangeQueryParameters = (newQueryParameters) => {
+      const newConfig = { ...eventDefinition.config, query_parameters: newQueryParameters || [] };
+
+      return onChange('config', newConfig);
+    };
+
+    if (!userCanViewLookupTables()) {
+      return (
+        <Alert bsStyle="info">
+          This account lacks permission to declare Query Parameters from Lookup Tables.
+        </Alert>
+      );
+    }
+
+    const parameterButtons = queryParameters.map((queryParam) => (
+      <EditQueryParameterModal key={queryParam.name}
+                               queryParameter={LookupTableParameter.fromJSON(queryParam)}
+                               embryonic={!!queryParam.embryonic}
+                               queryParameters={queryParameters}
+                               lookupTables={Object.values(tables)}
+                               onChange={onChangeQueryParameters} />
+    ));
+
+    if (isEmpty(parameterButtons)) {
+      return null;
+    }
+
+    const hasEmbryonicParameters = !isEmpty(queryParameters.filter((param) => (param.embryonic)));
+
+    return (
+      <FormGroup validationState={validation.errors.query_parameters ? 'error' : null}>
+        <ControlLabel>Query Parameters</ControlLabel>
+        <Alert bsStyle={hasEmbryonicParameters ? 'danger' : 'info'}>
+          <ButtonToolbar>
+            {parameterButtons}
+          </ButtonToolbar>
+        </Alert>
+        {hasEmbryonicParameters && (
+          <HelpBlock>
+            {validation.errors.query_parameters
+              ? get(validation, 'errors.query_parameters[0]')
+              : 'Please declare missing query parameters by clicking on the buttons above.'}
+          </HelpBlock>
+        )}
+      </FormGroup>
+    );
+  };
+
+  const onlyFilters = eventDefinition._scope === 'ILLUMINATE';
+
+  // Ensure deleted streams are still displayed in select
+  const allStreamIds = union(streams.map((s) => s.id), defaultTo(eventDefinition.config.streams, []));
+  const formattedStreams = formatStreamIds(allStreamIds);
+
+  return (
+    <fieldset>
+      <h2 className={commonStyles.title}>Filter</h2>
+      <p>Add information to filter the log messages that are relevant for this Event Definition.</p>
+      {onlyFilters || (
+      <Input id="filter-query"
+             name="query"
+             label="Search Query"
+             type="text"
+             help={(
+               <span>
+                 Search query that Messages should match. You can use the same syntax as in the Search page,
+                 including declaring Query Parameters from Lookup Tables by using the <code>$newParameter$</code> syntax.
+               </span>
+                 )}
+             value={defaultTo(eventDefinition.config.query, '')}
+             onChange={handleQueryChange} />
+      )}
+
+      {onlyFilters || renderQueryParameters()}
+
+      {!searchFiltersHidden && (
+      <FormGroup controlId="search-filters">
+        <ControlLabel>Search Filters <small className="text-muted">(Optional)</small></ControlLabel>
+        <div style={{ margin: '16px 0' }}>
+          <SearchFiltersFormControls filters={eventDefinition.config.filters}
+                                     onChange={handleSearchFiltersChange}
+                                     hideFiltersPreview={hideFiltersPreview} />
+        </div>
+      </FormGroup>
+      )}
+
+      {onlyFilters || (
+      <>
+        <FormGroup controlId="filter-streams">
+          <ControlLabel>Streams <small className="text-muted">(Optional)</small></ControlLabel>
+          <MultiSelect id="filter-streams"
+                       matchProp="label"
+                       onChange={(selected) => handleStreamsChange(selected === '' ? [] : selected.split(','))}
+                       options={formattedStreams}
+                       value={defaultTo(eventDefinition.config.streams, []).join(',')} />
+          <HelpBlock>Select streams the search should include. Searches in all streams if empty.</HelpBlock>
+        </FormGroup>
+
+        <FormGroup controlId="search-within" validationState={validation.errors.search_within_ms ? 'error' : null}>
+          <TimeUnitInput label="Search within the last"
+                         update={handleTimeRangeChange('search_within_ms')}
+                         value={searchWithinMsDuration}
+                         unit={searchWithinMsUnit}
+                         units={TIME_UNITS}
+                         clearable
+                         required />
+          {validation.errors.search_within_ms && (
+          <HelpBlock>{get(validation, 'errors.search_within_ms[0]')}</HelpBlock>
+          )}
+        </FormGroup>
+
+        <FormGroup controlId="execute-every" validationState={validation.errors.execute_every_ms ? 'error' : null}>
+          <TimeUnitInput label="Execute search every"
+                         update={handleTimeRangeChange('execute_every_ms')}
+                         value={executeEveryMsDuration}
+                         unit={executeEveryMsUnit}
+                         units={TIME_UNITS}
+                         clearable
+                         required />
+          {validation.errors.execute_every_ms && (
+          <HelpBlock>{get(validation, 'errors.execute_every_ms[0]')}</HelpBlock>
+          )}
+        </FormGroup>
+        <Input id="schedule-checkbox"
+               type="checkbox"
+               name="_is_scheduled"
+               label="Enable"
+               help="Should this event definition be executed automatically?"
+               checked={defaultTo(eventDefinition.config._is_scheduled, true)}
+               onChange={handleConfigChange} />
+      </>
+      )}
+    </fieldset>
+  );
+};
+
+export default FilterForm;
+
+FilterForm.propTypes = {
+  eventDefinition: PropTypes.object.isRequired,
+  validation: PropTypes.object.isRequired,
+  streams: PropTypes.array.isRequired,
+  onChange: PropTypes.func.isRequired,
+  currentUser: PropTypes.object.isRequired,
+};
